+++ lib_substr/parall_substr/parall_substr.c
@@ -5,46 +5,7 @@
 #include <sys/types.h>
 #include <sys/wait.h>
 #include <unistd.h>
-
 #include "parall_substr.h"
-
-#define ALPHABET_SIZE 256
-
-chunk_info* get_substr_info(const char* const str) {
-    size_t* count = (size_t*)calloc(ALPHABET_SIZE, sizeof(size_t));
-    chunk_info* chunk = (chunk_info*)calloc(1, sizeof(chunk_info));
-    if (!str || !count || !chunk) {
-        free(count);
-        free(chunk);
-        return NULL;
-    }
-    chunk->size = strlen(str);
-    for (size_t i = 0, lower_bound = 0,
-        len = strlen(str); i < len; i++) {
-        if (++count[str[i]] > 1) {
-            for (size_t j = lower_bound; j < i; j++) {
-                if (str[i] == str[j]) {
-                    lower_bound = j + 1;
-                    count[str[i]]--;
-                    break;
-                }
-            }
-        }
-        if (i == len - 1) {
-            chunk->max_r = len - lower_bound;
-        }  // правая граница
-        size_t tmp = i - lower_bound + 1;
-        if (tmp > chunk->max) {
-            chunk->max = tmp;
-            chunk->pos = lower_bound;
-            if (!lower_bound) {
-                chunk->max_l = tmp;
-            }  // левая граница
-        }
-    }
-    free(count);
-    return chunk;
-}
 
 chunk_info* merge_chunk_info(const chunk_info* const chunks, size_t num_chunks, FILE* const f) {
     if (!f || !chunks || !num_chunks) {
@@ -53,7 +14,7 @@ chunk_info* merge_chunk_info(const chunk_info* const chunks, size_t num_chunks, 
     rewind(f);
     chunk_info* merged_ch = (chunk_info*)calloc(1, sizeof(chunk_info));
     if (!merged_ch) {
-    	return NULL;
+        return NULL;
     }
     *merged_ch = *chunks;
     for (size_t i = 1; i < num_chunks; i++) {
@@ -67,15 +28,15 @@ chunk_info* merge_chunk_info(const chunk_info* const chunks, size_t num_chunks, 
             fseek(f, tmp_glob_pos, SEEK_SET);
             char* str = (char*)calloc(tmp_sz + 1, sizeof(char));
             if (!str) {
-            	free(merged_ch);
-            	return NULL;
+                free(merged_ch);
+                return NULL;
             }
             fgets(str, tmp_sz + 1, f);
             chunk_info* tmp = get_substr_info(str);
             if (!tmp) {
-            	free(str);
-            	free(merged_ch);
-            	return NULL;
+                free(str);
+                free(merged_ch);
+                return NULL;
             }
             if (merged_ch->max < tmp->max) {
                 merged_ch->max = tmp->max;
@@ -95,7 +56,7 @@ int comp_chunk_info(const chunk_info* l, const chunk_info* r) {
     return l->global_pos > r->global_pos;
 }
 
-char* get_max_substr_without_repeat_from_file(const char* const file_name) {
+char* get_max_substr_without_repeat_from_file_parall(const char* const file_name) {
     if (!file_name) {
         return NULL;
     }
@@ -130,7 +91,7 @@ char* get_max_substr_without_repeat_from_file(const char* const file_name) {
             free(tmp);
             exit(EXIT_SUCCESS);
         } else if (pid == -1) {
-        	exit(EXIT_FAILURE);
+            exit(EXIT_FAILURE);
         }
         glob_pos += len;
     }
@@ -139,12 +100,12 @@ char* get_max_substr_without_repeat_from_file(const char* const file_name) {
         int status = 0;
         wait(&status);
         if (!WIFEXITED(status)) {
-        	exit(EXIT_FAILURE);
+            exit(EXIT_FAILURE);
         }
     }
     chunk_info* chunk = (chunk_info*)calloc(num_proc, sizeof(chunk_info));
     if (!chunk) {
-    	fclose(f);
+        fclose(f);
         return NULL;
     }
     for (size_t i = 0; i < num_proc; i++) {
