+++ lib_substr/substr/substr.h
@@ -1,7 +1,9 @@
-#ifndef PROJECT_PARALL_SUBSTR_PARALL_SUBSTR_H_
-#define PROJECT_PARALL_SUBSTR_PARALL_SUBSTR_H_
+#ifndef PROJECT_SUBSTR_SUBSTR_H_
+#define PROJECT_SUBSTR_SUBSTR_H_
 
 #include <stddef.h>
+
+#define ALPHABET_SIZE 256
 
 typedef struct {
     size_t global_pos;  // абсолютная позиция чанка в строке
@@ -13,7 +15,6 @@ typedef struct {
 } chunk_info;
 
 chunk_info* get_substr_info(const char* const str);
-chunk_info* merge_chunk_info(const chunk_info* const chunks, size_t num_chunks, FILE* const f);
 char* get_max_substr_without_repeat_from_file(const char* const file_name);
 
-#endif  // PROJECT_PARALL_SUBSTR_PARALL_SUBSTR_H_
+#endif  // PROJECT_SUBSTR_SUBSTR_H_
